Now that common elements are factored out of BFT... ... and the tests are segregated, it makes no sense to split BFT tests so unify into a single test module for BFT.
