+++ test/test-digraph-algorithms.js
@@ -1,9 +1,10 @@
 // Encapsule/jsgraph/test/test-jsgraph-algorithms.js
 
+require('./test-digraph-algorithm-transpose');
+
 require('./test-digraph-algorithm-common-context');
 require('./test-digraph-algorithm-common-request');
 
-
-require('./test-digraph-algorithm-transpose');
 require('./test-digraph-algorithm-bft');
+
 require('./test-digraph-algorithm-dfsv');
